crawler/requirements.txt
queuelib
threaded
pyyaml
requests
influxdb-client